Rename "sdm-support" package "blueprint"
